backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq>=0.15.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
sse-starlette==2.1.3
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic[email]
aiosqlite>=0.19.0
